+++ media/lua/client/CHC_config.lua
@@ -36,6 +36,7 @@ CHC_settings = {
 
 local init_cfg = {
     show_recipe_module = true,
+    show_fav_items_inventory = false,
     recipe_selector_modifier = 1, -- none
     category_selector_modifier = 1,
     tab_selector_modifier = 1,
@@ -59,6 +60,7 @@ local init_cfg = {
 
 local function onModOptionsApply(values)
     CHC_settings.config.show_recipe_module = values.settings.options.show_recipe_module
+    CHC_settings.config.show_fav_items_inventory = values.settings.options.show_fav_items_inventory
     CHC_settings.config.recipe_selector_modifier = values.settings.options.recipe_selector_modifier
     CHC_settings.config.category_selector_modifier = values.settings.options.category_selector_modifier
     CHC_settings.config.tab_selector_modifier = values.settings.options.tab_selector_modifier
@@ -148,6 +150,13 @@ if ModOptions and ModOptions.getInstance then
                 default = true,
                 OnApplyMainMenu = onModOptionsApply,
                 OnApplyInGame = onModOptionsApply
+            },
+            show_fav_items_inventory = {
+                name = "IGUI_ShowFavItemsInventory",
+                tooltip = "IGUI_ShowFavItemsInventoryTooltip",
+                default = false,
+                OnApplyMainMenu = onModOptionsApply,
+                OnApplyInGame = onModOptionsApply
             }
         },
         mod_id = "CraftHelperContinued",
@@ -165,6 +174,7 @@ if ModOptions and ModOptions.getInstance then
 else
     -- defaults in case "Mod Options" not installed
     CHC_settings.config.show_recipe_module = true
+    CHC_settings.config.show_fav_items_inventory = false
     CHC_settings.config.recipe_selector_modifier = 1
     CHC_settings.config.category_selector_modifier = 1
     CHC_settings.config.tab_selector_modifier = 1